feat: move block exceptions to a proper folder (#181)
