+++ tests/Unit/Blocks/ColumnListTest.php
@@ -7,7 +7,7 @@ use Notion\Blocks\Column;
 use Notion\Blocks\ColumnList;
 use Notion\Exceptions\BlockException;
 use Notion\Blocks\Paragraph;
-use Notion\Exceptions\ColumnListException;
+use Notion\Exceptions\BlockException\ColumnListException;
 use PHPUnit\Framework\TestCase;
 
 class ColumnListTest extends TestCase
